version 1.9.1 GPS bug Update
-Intermec GPS – no longer needs the location services cab file, works
with a MS wrapper for GPS within the project code. GPS now works on
Windows Mobile 5.X devices.
-Datagrid highlighting – On the ATID, the back color is now blue instead
of black to help make the text more visible in sunlight situations.
On the Intermec, the Windows Mobile styling over-rides the color so it
will be unchanged, Windows Mobile 5.X = Blue, Windows Mobile 6.X =
Green….likely this could be changed in an OS setting / desktop theme. I
can actually control it separately if I wanted to but it takes some
trial and error, as blue on the ATID results in black on the Intermec,
so I think we can wait and see if any more user stories develop around
this, especially for the Intermec before we make further changes there.
-Add Address in the Database menu is now Add Service. Adhoc service
types in the drop down list used to be “Delivery” “Removal” are now
“DEL” and “REM”